requirements.txt
absl-py==1.4.0
astunparse==1.6.3
autopep8==2.0.1
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==2.10.0
idna==3.4
importlib-metadata==6.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.2.2
opencv-python==4.2.0.34
opt-einsum==3.3.0
Pillow==8.2.0
pipdeptree==2.5.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==2.2.0
tomli==2.0.1
tqdm==4.46.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
